Moved typesafe package outside of the model
