+++ src/main/scala/io/epifab/yadl/typesafe/Alias.scala
@@ -1,4 +1,4 @@
-package io.epifab.yadl.domain.typesafe
+package io.epifab.yadl.typesafe
 
 trait Alias[A]
