+++ src/main/scala/io/epifab/yadl/typesafe/DataSource.scala
@@ -1,6 +1,6 @@
-package io.epifab.yadl.domain.typesafe
+package io.epifab.yadl.typesafe
 
-import io.epifab.yadl.domain.typesafe.fields.{FieldDecoder, FieldEncoder}
+import io.epifab.yadl.typesafe.fields.{FieldDecoder, FieldEncoder}
 import io.epifab.yadl.utils.Appender
 import shapeless.{::, HList, HNil}
