+++ src/main/scala/io/epifab/yadl/typesafe/package.scala
@@ -1,4 +1,4 @@
-package io.epifab.yadl.domain
+package io.epifab.yadl
 
 package object typesafe {
   type AS[T, A] = T with Alias[A]
